# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OperationsStatus": {
        "properties": {
          "active_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunView"
              },
              {
                "type": "null"
              }
            ]
          },
          "api_status": {
            "title": "Api Status",
            "type": "string"
          },
          "api_version": {
            "title": "Api Version",
            "type": "string"
          },
          "can_refresh": {
            "title": "Can Refresh",
            "type": "boolean"
          },
          "last_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunView"
              },
              {
                "type": "null"
              }
            ]
          },
          "last_successful_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RunView"
              },
              {
                "type": "null"
              }
            ]
          },
          "refresh_configured": {
            "title": "Refresh Configured",
            "type": "boolean"
          },
          "serving": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ServingView"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "api_status",
          "api_version",
          "serving",
          "active_run",
          "last_run",
          "last_successful_run",
          "refresh_configured",
          "can_refresh"
        ],
        "title": "OperationsStatus",
        "type": "object"
      },
      "RefreshAccepted": {
        "properties": {
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          }
        },
        "required": [
          "run_id",
          "status"
        ],
        "title": "RefreshAccepted",
        "type": "object"
      },
      "RunStatus": {
        "enum": [
          "QUEUED",
          "RUNNING",
          "SUCCEEDED",
          "FAILED",
          "CANCELLED"
        ],
        "title": "RunStatus",
        "type": "string"
      },
      "RunView": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "current_stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Stage"
              },
              {
                "type": "null"
              }
            ]
          },
          "data_as_of": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data As Of"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "error_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Summary"
          },
          "github_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Run Id"
          },
          "github_run_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Run Url"
          },
          "job_type": {
            "title": "Job Type",
            "type": "string"
          },
          "requested_at": {
            "format": "date-time",
            "title": "Requested At",
            "type": "string"
          },
          "requested_by": {
            "title": "Requested By",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "serving_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serving Version"
          },
          "snapshot_outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SnapshotOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "stages": {
            "items": {
              "$ref": "#/components/schemas/StageRecord"
            },
            "title": "Stages",
            "type": "array"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "trigger": {
            "title": "Trigger",
            "type": "string"
          },
          "weekly_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Weekly Version"
          }
        },
        "required": [
          "run_id",
          "job_type",
          "trigger",
          "requested_by",
          "requested_at",
          "started_at",
          "completed_at",
          "duration_seconds",
          "status",
          "current_stage",
          "data_as_of",
          "weekly_version",
          "serving_version",
          "snapshot_outcome",
          "error_summary",
          "github_run_id",
          "github_run_url",
          "stages"
        ],
        "title": "RunView",
        "type": "object"
      },
      "RunsPage": {
        "properties": {
          "next_before": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Before"
          },
          "runs": {
            "items": {
              "$ref": "#/components/schemas/RunView"
            },
            "title": "Runs",
            "type": "array"
          }
        },
        "required": [
          "runs",
          "next_before"
        ],
        "title": "RunsPage",
        "type": "object"
      },
...
      "SnapshotOutcome": {
        "enum": [
          "PUBLISHED",
          "UNCHANGED",
          "NOT_PUBLISHED"
        ],
        "title": "SnapshotOutcome",
        "type": "string"
      },
      "SnapshotView": {
        "properties": {
          "analysis": {
            "anyOf": [
              {
                "additionalProperties": {
                  "anyOf": [
                    {
                      "type": "string"
                    },
                    {
                      "type": "integer"
                    }
                  ]
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis"
          },
          "counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "title": "Counts",
            "type": "object"
          },
          "data_as_of": {
            "format": "date",
            "title": "Data As Of",
            "type": "string"
          },
          "live": {
            "title": "Live",
            "type": "boolean"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "schema_version": {
            "title": "Schema Version",
            "type": "integer"
          },
          "snapshot_id": {
            "title": "Snapshot Id",
            "type": "string"
          },
          "staged_at": {
            "format": "date-time",
            "title": "Staged At",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "versions": {
            "additionalProperties": {
              "type": "string"
            },
            "title": "Versions",
            "type": "object"
          }
        },
        "required": [
          "snapshot_id",
          "live",
          "status",
          "staged_at",
          "published_at",
          "run_id",
          "data_as_of",
          "schema_version",
          "versions",
          "counts"
        ],
        "title": "SnapshotView",
        "type": "object"
      },
      "Stage": {
        "enum": [
          "INGEST",
          "CORPORATE_ACTIONS",
          "ADJUSTMENT",
          "DATA_QUALITY",
          "WEEKLY",
          "ANALYSIS",
          "PUBLISH_SERVING"
        ],
        "title": "Stage",
        "type": "string"
      },
      "StageRecord": {
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "details": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "number"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ]
            },
            "title": "Details",
            "type": "object"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "error_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Summary"
          },
          "records_processed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Records Processed"
          },
          "stage": {
            "$ref": "#/components/schemas/Stage"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus",
            "default": "QUEUED"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "stage"
        ],
        "title": "StageRecord",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "See the structure. Read the trend.",
    "title": "ChartLens API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/jobs": {
      "get": {
        "operationId": "list_jobs_api_v1_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "requested_at cursor",
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "requested_at cursor",
              "title": "Before"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunsPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Jobs",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/v1/jobs/{run_id}": {
      "get": {
        "operationId": "get_job_api_v1_jobs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "maxLength": 80,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Job",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/v1/jobs/{run_id}/cancel": {
      "post": {
        "description": "Cancel a refresh that no workflow has started (admins only). It frees the lock if\nGitHub dropped the workflow before it ran. A running refresh is cancelled in GitHub\nActions, where the workflow records it.",
        "operationId": "cancel_job_api_v1_jobs__run_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "maxLength": 80,
              "pattern": "^[A-Za-z0-9][A-Za-z0-9_.-]{0,79}$",
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Job",
        "tags": [
          "operations"
        ]
      }
    },
...
    "/api/v1/refresh/daily": {
      "post": {
        "description": "Start the production refresh. Admins only. Returns at once with the run id: the\nrefresh runs in GitHub Actions, and ``GET /jobs/{run_id}`` follows it.",
        "operationId": "refresh_daily_api_v1_refresh_daily_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefreshAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "A refresh is already active"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Daily",
        "tags": [
          "operations"
        ]
      }
    },
...
    "/api/v1/system/operations": {
      "get": {
        "operationId": "operations_status_api_v1_system_operations_get",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationsStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Operations Status",
        "tags": [
          "operations"
        ]
      }
    },
    "/api/v1/system/snapshots": {
      "get": {
        "description": "Published snapshot history, newest first. The API still serves only what the\npointer names; this is a record of publications, never a way to choose one.",
        "operationId": "list_snapshots_api_v1_system_snapshots_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SnapshotView"
                  },
                  "title": "Response List Snapshots Api V1 System Snapshots Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Snapshots",
        "tags": [
          "operations"
        ]
      }
    },
...
  }
}